refactored tables creation into services/tables.js
