+++ footnote-frontend/src/main.tsx
@@ -1,14 +1,11 @@
 import { StrictMode } from "react";
 import { createRoot } from "react-dom/client";
-import { BrowserRouter } from "react-router-dom"; // Import BrowserRouter
 import App from "./App.tsx";
 
 createRoot(document.getElementById("root")!).render(
   <StrictMode>
-    <BrowserRouter>
       {" "}
       {/* Wrap App with BrowserRouter */}
       <App />
-    </BrowserRouter>
   </StrictMode>
 );
